schema-registry service renamed to registry
