+++ examples/jsonplaceholder/docker-compose.yml
@@ -5,14 +5,14 @@ services:
     build:
       context: ../../services/graphql-server
     depends_on:
-      - graphql-gateway.schema-registry
+      - graphql-gateway.registry
     healthcheck:
       disable: true
     ports:
       - 3000:8011
-  graphql-gateway.schema-registry:
+  graphql-gateway.registry:
     build:
-      context: ../../services/schema-registry
+      context: ../../services/registry
       args:
         target: PRODUCTION
     healthcheck:
