+++ examples/kubernetes/skaffold.yaml
@@ -6,8 +6,8 @@ build:
     context: ../../services/graphql-server
     docker:
       dockerfile: dockerfile
-  - image: schema-registry
-    context: ../../services/schema-registry
+  - image: registry
+    context: ../../services/registry
     docker:
       dockerfile: dockerfile
   - image: gql-controller
@@ -20,7 +20,7 @@ deploy:
     - ./deployments/infra/rbac.yaml
 
     - ./deployments/infra/admission-tls-secret.yaml
-    - ./deployments/infra/schema-registry.yaml
+    - ./deployments/infra/registry.yaml
     - ./deployments/infra/gql-controller.yaml
     - ./deployments/infra/gateway.yaml
