+++ examples/user-subscriptions/docker-compose.override.yaml
@@ -5,15 +5,15 @@ services:
     image: byrnedo/alpine-curl
     volumes:
       - ./user-service:/data
-    command: -X POST --header "Content-Type:text/plain" -d @/data/schema.gql http://graphql-gateway.schema-registry/schema/user
+    command: -X POST --header "Content-Type:text/plain" -d @/data/schema.gql http://graphql-gateway.registry/schema/user
     depends_on:
-      graphql-gateway.schema-registry:
+      graphql-gateway.registry:
         condition: service_healthy
   user-subscription-registration:
     image: byrnedo/alpine-curl
     volumes:
       - ./user-subscription-service:/data
-    command: -X POST --header "Content-Type:text/plain" -d @/data/schema.gql http://graphql-gateway.schema-registry/schema/user-subscription
+    command: -X POST --header "Content-Type:text/plain" -d @/data/schema.gql http://graphql-gateway.registry/schema/user-subscription
     depends_on:
-      graphql-gateway.schema-registry:
+      graphql-gateway.registry:
         condition: service_healthy
