+++ services/registry/.dockerignore
