+++ services/registry/Dockerfile
