+++ services/registry/jest.config.js
