+++ services/registry/src/graphql/create-schema.ts
