+++ services/registry/src/graphql/schema-utils.test.ts
