+++ services/registry/src/graphql/schema-utils.ts
