+++ services/registry/src/index.ts
