+++ services/registry/src/routes/schema.ts
