+++ services/registry/src/routes/validate.ts
