+++ services/registry/src/sources-config.ts
