+++ services/registry/src/sources/remote-source-template/index.ts
