+++ services/registry/src/sources/remote-source-template/remote-source.ts
