+++ services/registry/src/sources/remote-source-template/source-type.ts
