+++ services/registry/src/sync/sync-service.ts
