+++ services/registry/tsconfig.json
